refactor: extract email_template http feature into ferriskey-api-email-template
