+++ api/Cargo.toml
@@ -14,6 +14,7 @@ ferriskey-api-compass = { path = "../libs/ferriskey-api-compass" }
 ferriskey-api-webhook = { path = "../libs/ferriskey-api-webhook" }
 ferriskey-api-health = { path = "../libs/ferriskey-api-health" }
 ferriskey-api-role = { path = "../libs/ferriskey-api-role" }
+ferriskey-api-email-template = { path = "../libs/ferriskey-api-email-template" }
 anyhow = "1.0.97"
 axum = "0.8.8"
 axum-cookie = "0.2.3"
