+++ libs/ferriskey-api-email-template/src/handlers/mod.rs
