+++ libs/ferriskey-api-email-template/src/handlers/update_template.rs
@@ -6,60 +6,66 @@ use ferriskey_core::domain::{
     authentication::value_objects::Identity,
     email_template::{
         entities::EmailTemplate,
-        ports::{EmailTemplateService, GetEmailTemplateInput},
+        ports::{EmailTemplateService, UpdateEmailTemplateInput},
     },
 };
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 use uuid::Uuid;
 
-use crate::application::http::server::{
-    api_entities::{
-        api_error::{ApiError, ApiErrorResponse},
-        response::Response,
-    },
-    app_state::AppState,
+use crate::validators::UpdateEmailTemplateValidator;
+use ferriskey_api_core::api_entities::{
+    api_error::{ApiError, ApiErrorResponse, ValidateJson},
+    response::Response,
 };
+use ferriskey_api_core::app_state::AppState;
 
 #[derive(Debug, Serialize, Deserialize, ToSchema, PartialEq)]
-pub struct GetEmailTemplateResponse {
+pub struct UpdateEmailTemplateResponse {
     pub data: EmailTemplate,
 }
 
 #[utoipa::path(
-    get,
+    put,
     path = "/{template_id}",
     tag = "email-template",
-    summary = "Get email template",
-    description = "Retrieves a single email template by ID.",
+    summary = "Update email template",
+    description = "Updates an existing email template.",
     params(
         ("realm_name" = String, Path, description = "Name of the realm"),
         ("template_id" = Uuid, Path, description = "Email template ID"),
     ),
+    request_body = UpdateEmailTemplateValidator,
     responses(
-        (status = 200, description = "Email template retrieved successfully", body = GetEmailTemplateResponse),
+        (status = 200, description = "Email template updated successfully", body = UpdateEmailTemplateResponse),
+        (status = 400, description = "Invalid request data", body = ApiErrorResponse),
         (status = 404, description = "Email template not found", body = ApiErrorResponse),
         (status = 401, description = "Unauthorized", body = ApiErrorResponse),
         (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
         (status = 500, description = "Internal server error", body = ApiErrorResponse),
     ),
 )]
-pub async fn get_template(
+pub async fn update_template(
     Path((realm_name, template_id)): Path<(String, Uuid)>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
-) -> Result<Response<GetEmailTemplateResponse>, ApiError> {
+    ValidateJson(payload): ValidateJson<UpdateEmailTemplateValidator>,
+) -> Result<Response<UpdateEmailTemplateResponse>, ApiError> {
     let template = state
         .service
-        .get_template(
+        .update_template(
             identity,
-            GetEmailTemplateInput {
+            UpdateEmailTemplateInput {
                 realm_name,
                 template_id,
+                name: payload.name,
+                structure: payload.structure,
             },
         )
         .await
         .map_err(ApiError::from)?;
 
-    Ok(Response::OK(GetEmailTemplateResponse { data: template }))
+    Ok(Response::Updated(UpdateEmailTemplateResponse {
+        data: template,
+    }))
 }
