+++ libs/ferriskey-api-email-template/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod handlers;
+pub mod router;
+pub mod validators;
