+++ libs/ferriskey-api-email-template/src/validators.rs
